separated hardware and load logic into separate files, added load shedding
